% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm.R
\name{generate_mirrors}
\alias{generate_mirrors}
\title{Generate noise variable for mirrored feature construction}
\usage{
generate_mirrors(y, x, family, ncores, nlambda)
}
\arguments{
\item{y}{response variable. Quantitative for family="gaussian", or family="poisson" (non-negative counts). For family="binomial" should be either a factor with two levels.}

\item{x}{input design matrix of dimension n x p; each row is an observation vector.}

\item{family}{response type.}

\item{ncores}{number of cores used for parallel computation of the GM_1 to GM_p. We recommend to use multiple cores which will greatly reduce the computational time.}

\item{nlambda}{The number of lambda values. The default is 100.}
}
\value{
z: matrix containing noise variables Z at the right scale for constructing mirrors x + z, x - z
}
\description{
Generate noise variable for mirrored feature construction
}
