% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-gm-internal.R
\name{gm-internal}
\alias{gm-internal}
\alias{to_value}
\alias{to_value.Pitch}
\alias{to_value.Tupler}
\alias{to_value.Duration}
\alias{to_value.Meter}
\alias{add}
\alias{add.Clef}
\alias{add.Key}
\alias{add.Meter}
\alias{add.Tempo}
\alias{add.Line}
\alias{normalize_pitch}
\alias{normalize_pitch.numeric}
\alias{normalize_pitch.character}
\alias{normalize_pitch.default}
\alias{print.PitchLine}
\alias{print.PitchNotation}
\alias{print.PitchValue}
\alias{print.PitchRest}
\alias{print.PitchChord}
\alias{print.Pitch}
\alias{print.ClefLine}
\alias{print.DurationLine}
\alias{print.KeyLine}
\alias{print.MeterLine}
\alias{print.Move}
\alias{print.Note}
\alias{print.Rest}
\alias{print.Measure}
\alias{print.Attributes}
\alias{print.Score}
\alias{print.Element}
\alias{print.TempoLine}
\alias{print.PositionLine}
\alias{to_Pitch}
\alias{to_Pitch.default}
\alias{to_Pitch.numeric}
\alias{to_Pitch.character}
\alias{to_Pitch.PitchNotation}
\alias{to_Pitch.PitchValue}
\alias{to_Pitch.PitchChord}
\alias{to_Pitch.Line}
\alias{to_Pitch.Music}
\alias{to_Duration}
\alias{to_Duration.default}
\alias{to_Duration.Duration}
\alias{to_Duration.numeric}
\alias{to_Duration.character}
\alias{check_duration}
\alias{check_duration.numeric}
\alias{check_duration.character}
\alias{to_Element}
\alias{to_Element.PitchRest}
\alias{to_Element.Pitch}
\alias{to_Element.Clef}
\alias{to_Element.Key}
\alias{to_Element.Meter}
\alias{to_Element.Score}
\alias{to_Element.Part}
\alias{to_Element.Measure}
\alias{to_Element.Attributes}
\alias{to_Element.Move}
\alias{to_Element.Rest}
\alias{to_Element.Note}
\alias{to_Element.default}
\alias{to_Element.Tempo}
\alias{check_positions_length}
\alias{check_positions_length.numeric}
\alias{check_positions_content}
\alias{check_positions_content.numeric}
\alias{check_positions_content.list}
\alias{==.Pitch}
\alias{+.ClefLine}
\alias{+.KeyLine}
\alias{+.MeterLine}
\alias{+.TempoLine}
\alias{check_pitches}
\alias{check_pitches.default}
\alias{check_pitches.logical}
\alias{check_pitches.numeric}
\alias{check_pitches.character}
\alias{check_pitches.list}
\alias{signify}
\alias{signify.default}
\alias{signify.character}
\alias{signify.list}
\alias{signify.Pitch}
\alias{signify.Tupler}
\alias{signify.Meter}
\alias{signify.Key}
\alias{quantify}
\alias{quantify.default}
\alias{quantify.numeric}
\alias{quantify.list}
\alias{quantify.Pitch}
\alias{quantify.Tupler}
\alias{quantify.Duration}
\alias{quantify.Meter}
\alias{quantify.Key}
\alias{check_durations}
\alias{check_durations.default}
\alias{check_durations.numeric}
\alias{check_durations.character}
\alias{check_durations.list}
\alias{check_duration_length}
\alias{check_duration_length.numeric}
\alias{check_duration_length.character}
\alias{to_musicxml}
\alias{to_musicxml.Element}
\title{Internal gm Functions}
\description{
Internal gm functions.
These are not to be called by the user.
}
\keyword{internal}
