% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm.R
\name{gm}
\alias{gm}
\title{Gaussian Mirrors for Controlled Variable Selection}
\usage{
gm(y, x, q = 0.1, family = "gaussian", ncores = 1, rep.boot = 100,
  var.est = FALSE, nlambda = 100)
}
\arguments{
\item{y}{response variable. Quantitative for family="gaussian", or family="poisson" (non-negative counts). For family="binomial" should be either a factor with two levels.}

\item{x}{input design matrix of dimension n x p; each row is an observation vector.}

\item{q}{target FDR level. The default value is 0.1.}

\item{family}{response type.}

\item{ncores}{number of cores used for parallel computation of the GM_1 to GM_p. We recommend to use multiple cores which will greatly reduce the computational time.}

\item{rep.boot}{the size of bootstrap sample for estimating the variance of FDR.}

\item{var.est}{logical flag for estimating the variance of FDR. The default value is FALSE.}

\item{nlambda}{The number of lambda values. The default is 100.}
}
\value{
gm_selected: the index of selected variable.

gm_statistics: the value for mirror statistics.

gm_var: if the var.est=TRUE, return the estimated variance of FDR.
}
\description{
Gaussian Mirrors for Controlled Variable Selection
}
\examples{
\donttest{n <- 300  # Number of observations
p <- 1000  # Number of predictors included in model
q <- 60  # Number of true predictors
amplitude = 20 # signal amplitude (for noise level = 1) 


# Generate the variables from a multivariate normal distribution
times <- 1:p
sigma <- 1
rho = 0.5
H <- abs(outer(times, times, "-"))
V <- sigma * rho^H
x = mvrnorm(n, mu = rep(0,p), Sigma=V)

# Generate the response.
true_index= sample(p,q)
beta = numeric(p)
beta[true_index]=beta_q
mu = x \%*\%  beta
y = mu + rnorm(n)

# Fit the Gaussian mirror model
fit = gm(y, x, ncores=4)
}


}
\references{
\url{}
}
\author{
Xin Xing, \email{xin_xing@fas.harvard.edu}
}
\keyword{fdr}
\keyword{gaussian}
\keyword{mirror,}
\keyword{selection,}
\keyword{variable}
