% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{inspect_errors}
\alias{inspect_errors}
\title{See Full Error Report}
\usage{
inspect_errors()
}
\value{
Invisible \code{NULL}.
}
\description{
See a full error report when the error message is too
long and thus shortened.
}
\examples{
\dontrun{
Line(list(c, "p", NULL, 1:3, TRUE, NA_character_))
}
}
