% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\docType{package}
\name{mr}
\alias{mr}
\title{mr: Generate Musical Scores Easily and Show Them Anywhere}
\description{
mr provides a simple and intuitive language, with which you
can create musical scores and audio files easily. Also, mr works in R
Markdown documents, R Jupyter Notebooks and RStudio, so you can embed
scores and audio files anywhere. You only tell mr the structure of your
music, and it takes care of all the dirty technical details.
}
\section{Author}{

Renfei Mao \href{mailto:renfeimao@gmail.com}{renfeimao@gmail.com}
}

