% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R, R/show.R
\name{show}
\alias{show}
\alias{show.Music}
\title{Show Object}
\usage{
show(x, to)

\method{show}{Music}(x, to = NULL)
}
\arguments{
\item{x}{An object.}

\item{to}{Optional. A character vector which contains "score", "audio" or
both, which indicates whether to show the object as musical score or
audio file.}
}
\value{
Invisible \code{NULL}.
}
\description{
Show an object as musical score or audio file.
}
\section{Methods (by class)}{
\itemize{
\item \code{Music}: show a \code{Music} object.
}}

\examples{
m <- Music() + Meter(4, 4) + Line(list("C4"), list(4))
show(m, c("score", "audio"))
}
