% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomato.R
\docType{data}
\name{tomato}
\alias{tomato}
\title{GWAS data set with a panel of 292 tomato accessions}
\format{each column corresponding a SNP}
\usage{
data(tommato)
}
\description{
This  dataset  is  publicly  available at (\href{ftp://ftp.solgenomics.net/manuscripts/Bauchet_2016/})
}
\examples{
data(tommato)
}
\references{
Bauchet et al. (2017) Theoretical and applied genetics, 130(5):875–889.
}
\keyword{datasets}
