% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{tuplet}
\alias{tuplet}
\title{Create Tuplet}
\usage{
tuplet(duration, ...)
}
\arguments{
\item{duration}{A duration notation, duration value, or \code{Duration} object.}

\item{...}{\code{Tupler} objects returned by \code{\link[=Tupler]{Tupler()}}, which specify
how to divide the argument \code{duration} into parts, and how to take from
these parts.}
}
\value{
A list with class \code{Duration}.
}
\description{
Create a tuplet.
}
\examples{
# create a triplet quarter note
tuplet("half", Tupler(3, unit = "quarter", take = "quarter"))
}
\seealso{
\code{\link[=Tupler]{Tupler()}}

\code{vignette("mr", package = "mr")} for a friendly guide to tuplets.
}
