% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm-internal.R
\name{gm-internal}
\alias{gm-internal}
\alias{add}
\alias{add.Accidental}
\alias{add.Articulation}
\alias{add.Clef}
\alias{add.Dynamic}
\alias{add.Fermata}
\alias{add.Grace}
\alias{add.Hairpin}
\alias{add.Instrument}
\alias{add.Key}
\alias{add.Line}
\alias{add.Lyric}
\alias{add.Meter}
\alias{add.Mordent}
\alias{add.Schleifer}
\alias{add.Tremolo}
\alias{add.Turn}
\alias{add.Notehead}
\alias{add.Velocity}
\alias{add.Tie}
\alias{add.Breath}
\alias{add.Pedal}
\alias{add.Slur}
\alias{add.Stem}
\alias{add.Tempo}
\alias{locate}
\alias{locate.Accidental}
\alias{locate.Articulation}
\alias{locate.Clef}
\alias{locate.Dynamic}
\alias{locate.Fermata}
\alias{locate.Grace}
\alias{locate.Hairpin}
\alias{locate.Instrument}
\alias{locate.Key}
\alias{locate.Meter}
\alias{locate.Notehead}
\alias{locate.Velocity}
\alias{locate.Tie}
\alias{locate.Breath}
\alias{locate.Slur}
\alias{locate.Stem}
\alias{locate.Tempo}
\alias{check_add_to}
\alias{check_add_to.default}
\alias{check_add_to.character}
\alias{check_add_to.numeric}
\alias{to_string}
\alias{to_string.Clef}
\alias{to_string.Key}
\alias{to_string.Pitch}
\alias{to_string.Duration}
\alias{to_string.Meter}
\alias{to_string.MusicXML}
\alias{to_value}
\alias{to_value.Pitch}
\alias{to_value.Duration}
\alias{to_value.Meter}
\alias{to_Pitch}
\alias{to_Pitch.character}
\alias{to_Pitch.numeric}
\alias{to_Duration}
\alias{to_Duration.character}
\alias{to_MusicXML}
\alias{to_MusicXML.Pitch}
\alias{to_MusicXML.Breath}
\alias{to_MusicXML.Duration}
\alias{to_MusicXML.Note}
\alias{to_MusicXML.Tempo}
\alias{to_MusicXML.Attributes}
\alias{to_MusicXML.Music}
\alias{to_MusicXML.Key}
\alias{to_MusicXML.Meter}
\alias{to_MusicXML.Accidental}
\alias{to_MusicXML.Articulation}
\alias{to_MusicXML.Clef}
\alias{to_MusicXML.Dynamic}
\alias{to_MusicXML.Fermata}
\alias{to_MusicXML.Hairpin}
\alias{to_MusicXML.Pedal}
\alias{to_MusicXML.Notehead}
\alias{to_MusicXML.Stem}
\alias{to_MusicXML.Schleifer}
\alias{to_MusicXML.Mordent}
\alias{to_MusicXML.Slur}
\alias{to_MusicXML.Turn}
\alias{to_MusicXML.Trill}
\alias{to_MusicXML.Tremolo}
\alias{to_MusicXML.Lyric}
\alias{to_MusicXML.Instrument}
\alias{to_fraction}
\alias{to_fraction.Meter}
\alias{to_fraction.Duration}
\alias{to_fraction.numeric}
\alias{print.Pitch}
\alias{print.Duration}
\alias{print.MusicXML}
\alias{insert}
\alias{insert.Attributes}
\alias{insert.Key}
\alias{insert.Meter}
\alias{insert.Accidental}
\alias{insert.Articulation}
\alias{insert.Breath}
\alias{insert.Tempo}
\alias{insert.Clef}
\alias{insert.Dynamic}
\alias{insert.Fermata}
\alias{insert.Hairpin}
\alias{insert.Pedal}
\alias{insert.Notehead}
\alias{insert.Stem}
\alias{insert.Schleifer}
\alias{insert.Mordent}
\alias{insert.Slur}
\alias{insert.Turn}
\alias{insert.Trill}
\alias{insert.Tremolo}
\alias{insert.Lyric}
\alias{insert.Instrument}
\alias{export.MusicXML}
\alias{show.MusicXML}
\title{Internal gm Functions}
\description{
These are not to be called by the user.
}
\keyword{internal}
