% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm.R
\name{gm}
\alias{gm}
\title{Gaussian Mirrors for Controlled Variable Selection}
\usage{
gm(y, x, family = "gaussian", ncores = 1)
}
\arguments{
\item{y}{response variable}

\item{x}{predictor variables}

\item{family}{}

\item{q}{target FDR level}
}
\value{
result: the index of selected variable


}
\description{
Gaussian Mirrors for Controlled Variable Selection
}
\examples{
gm(y, x)

}
\references{
\url{}
}
\author{
Xin Xing, \email{xin_xing@stat.harvard.edu}
}
\keyword{selection}
\keyword{variable}
