% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm.R
\docType{package}
\name{gm}
\alias{gm}
\title{gm: Generate Musical Scores Easily and Show Them Anywhere}
\description{
Provides a simple and intuitive language, with which you can
create complex music easily. Takes care of all the dirty technical
details in converting your music to musical scores and audio files.
Works in R Markdown documents, R Jupyter Notebooks and RStudio, so you
can embed musical scores and audio files anywhere.
}
\section{Author}{

Renfei Mao \href{mailto:renfeimao@gmail.com}{renfeimao@gmail.com}
}

