% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm.R
\docType{package}
\name{gm}
\alias{gm}
\title{gm: Create Music and Show Them Anywhere}
\description{
Provides a simple and intuitive high-level language,
with which you can create  music easily. Takes care of all the dirty
technical details in converting your music to musical scores and audio
files. Works in R Markdown documents, R Jupyter Notebooks and RStudio,
so you can embed generated music anywhere.
}
\section{Author}{

Renfei Mao \href{mailto:renfeimao@gmail.com}{renfeimao@gmail.com}
}

