% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-construct-music.R
\name{Music}
\alias{Music}
\title{Initialize \code{Music} Object}
\usage{
Music()
}
\value{
A list of class \code{Music}.
}
\description{
Initialize a \code{Music} object. Other components can be added to it.
}
\examples{
# Initialize a `Music`
Music()
}
\seealso{
\code{\link[=+.Music]{+.Music()}} for adding components to a \code{Music} object.
}
